+++ changelog.md
@@ -1,6 +1,16 @@
 CHANGELOG
 ---------
 
+##1.1.3
+#### Feature
+- added `tag` settings to text/textarea to let user choose html tag
+- added `public` settings to control entity variable visibility
+- added `sizes` in image and gallery
+- leave only fallback for map as it is now included in acf
+
+#### Bugfix
+- Fix post selection when using term template
+
 ##1.1.2
 #### Feature
 - added `term template` display rule selector, require a `get_taxonomy_templates` function